requirements.txt
# pytorch
pytorch-lightning
omegaconf
wandb
numpy
pandas
rich
matplotlib
torchmetrics[image]
